Fix Issue 36.
lines() function is now complete and can be used for both reading and writing. For reading lines of text containing only a single column of data, it takes an additional argument as {convert:schema} where one can specify the schema type, and lines() will return an array of the specified type.

write(lines()) now produces a text file.

Added entry for lines() in storage-default.jql

Added tests for the new function.