Removed toStrings and array copies if possible for the conversion from JSON to del (delimited) and improved its performance.
Added symmetric handling of quoted and escape options for the conversion between 
JSON and del. 

For the conversion from JSON to del, quoted and escape does not apply to 
JSONBool, JSONLong, JSONDecimal and JSONDouble. Values of these types are always 
written out as is.

1. quoted: If false, JSON values are written out as is. For example, a"c is
written out as a"c. If true, prefixing and suffixed double quotes are added. And
one double quote is escaped into 2 double quotes. For example, a"c are
converted into "a""c".
2. escape: Only effective if quoted is true. If false, no backslash escaping. For
example, \n (1 character) is written as \n (1 character). If true, backslash
escaping happens. For example, \n (1 character) is written out as \\n (2
characters). And ISO control character is escaped into an unicode escape
sequence in the format of \uxxxx.

For the conversion from del to JSON: 

1. quoted: If quoted is false, del fields are set to JSON value as is. If quoted
is true, unquoting and double quote unescaping happen only if the del fields
has prefixing and suffixing double quotes. For example, 122 is set to a JSON
value as is. "ab""c" is converted into ab"c.

2. escape: effective only if quoted is true. If escape is true, backslash
unescaping happens. For example, \\n (2 character) is converted to \n (one
characer). \u0001 is converted into the character represented by the unicode
sequence. \" , \h and \u1xx2 are considered illegal. They are just set to JSON
values as is.