First cut at adding more control over exception handling. The two new expressions are: 1) catch(e, {errThresh: long}? opts);
 Eval e. if more than errThresh exceptions occur, abort. This lets us skip over some number of faulty inputs. if opts is not specified, then a global policy is used. It can be set using the following:
 
2) registerExceptionHandler({errThresh: long}?);

Note that the global policy has state, namely the number of exceptions that have occurred. This is state is
local to a JVM. When the threshold is specified in the catch function, then it is local to that call-site.

See failQueries.txt for further examples.

Future work here will allow exception handling to be further customized.